refactor: extract authentication logic into separate function for cleaner code
